Address review feedback
- Change mode format to %#04o to always get a leading 0 and a minimum of 4 digits.
- Change Stat of Lstat in assertSameFile
- Reduce code duplication between linux and bsd platforms.
